Use unsigned numbers for do "indexors" in t-routine.c
The upcoming name for an index that can be a THROWN_FLAG or an END_FLAG
as well as an unsigned REBCNT is "indexor" (which could simply be read
as index-or-a-flag, but it will also have some XOR 0 magic used for
another iteration mode).  A signed number cannot be used for them.

This updates two cases where a signed int was used to hold the result
of a DO/DO_NEXT operation, from %t-routine.c