Fix molding of typesets
A debug modification was committed that was desired to modify the
presentation of typesets when they are resident in the key slots of
contexts or function parameter lists.  These should not be exposed
to the user, so their molding behavior should be in service to the
debug dump (which wants to see the symbols of a range of typesets
vs. usually needing to see the types).

This modification was supposed to only be triggered when the typeset
symbol was not SYM_0, but a portion of it was outside the SYM_0 check
causing it to apply to ordinary typesets as well.  This fixes that,
as well as the initialization of the default typesets to set the
symbol to SYM_0 instead of leaving it as uninitialized bits.