Distinguish opt-any-value! from any-value!
In the rethinking of how optional parameters would work--along with
the general role of "unset!" within the system--the idea came to be
that UNSET! is not a "value" but the absence of a value.  To truly
have a "meta" type that can be queried for the full range of values
that are "in-band" and indicate their absence, one type must be
removed...and UNSET! seemed to be that.

What was formerly ANY-TYPE! was to become ANY-VALUE! and then
optionality be signaled by <opt> on the parameter.  However, debates
went on about whether the signals would be TAG! or ISSUE! in the
function generators, as well as what level the "reification" of unsets
would be prohibited.

This has held up other things, so this commit takes an intermediary
step of accomplishing `<opt> any-type!` with a typeset opt-any-type!.
It also adds ANY-NOTHING! and ANY-SOMETHING! for conveying whether a
value is a reified unset! (NONE!) or a non-reified one (UNSET!)