Add ENSURE mezzanine (with ENSURE/TYPE)
This adds a construct designed to pass through a value, assuming it
is not NONE! or UNSET!... e.g. that it is a SOMETHING?.  If it is a
NOTHING? then it will fail with an error message.

There is also a refinement that is able to take a data type or types
to check.  If types are provided to the ENSURE then they override the
checking for set and unset--only those types will match, and NONE!
or UNSET! will be fair game.

    >> ensure 1020
    == 1020

    >> ensure none
    ** user error: ENSURE did not expect value to have type none!

    ** Note: use WHY? for more error information

    >> ensure if 304 > 1020 [print "unsets aren't taken"]
    ** user error: ENSURE did not expect value to have type unset!

    >> ensure/type 1020 string!
    ** user error: ENSURE did not expect value to have type integer!

    >> ensure/type 1020 [string! integer!]
    == 1020

    >> ensure/type none none!
    == none