Make ASSERT return UNSET! instead of TRUE
The ASSERT function would return either TRUE or fail, which wasn't very
useful.  With the ENSURE function picking up the slack on "passthru"
situations for asserting, rather than enhance assert to do something
in the same vein this changes it to return unset for the sake of
safer error checking.