Improve UTF-8 decoding error messages
The message regarding a codepoint in UTF-8 being too high is more
likely to be triggered by the source data being not UTF-8 at all.  So
this guides people in that direction of inquiry with the error message
while still considering the possibility that it was UTF-8 with a big
codepoint.

While doing that, this observes that the generic `bad-decode` error
was really only being used to report UTF-8 error messages at this time.
Given the frequency of hitting this, switching it to a UTF-8 specific
message is likely to reduce the support burden for people who have
non-UTF-8 data they're trying to use with TO-STRING, etc.