Turn README.md into mostly a redirect page
Because the README.md is in the repository (and held also to the 80
column rule), it is not convenient to update like a wiki or Trello.
This reduces the amount of content in the file.

Ultimately it should have examples of how to call the library when
that is ready to publish.