More descriptive names for GET_VAR macros
The GET_VAR macros are not called all that often, but when they do get
called it is important to raise attention to what that means.  Given
that being able to return an unset is now expressed as GET/OPT, that
convention is extented to make GET_OPT_VAR to communicate that if
one isn't okay with UNSET!, then processing must be done.

Additionally since these operations can fail (and longjmp) if the word
is unbound, it's useful to point attention to that.  The TRY_GET_VAR
variants have "TRY" on them but the non-TRY versions now say MAY_FAIL.

By having more descriptive names, it becomes less necessary to make
comments at the callsites to point out the potential pitfalls...and
raises awareness of the possible outcomes.