Prefetching evaluator (full rewrite)
This is a top-to-bottom rewrite of the evaluator, which hones its ways
of acquiring and updating values to process into a stylized API that
allows Do_Core to serve as the basis for all(*) evaluation requests.

(* - "extension" callbacks are not yet using the method, but will.)

This recasts the evaluator to build upon three basic primitives for
navigating the input data (really two, but three for clarity):

    - FETCH_NEXT_ONLY_MAYBE_END
    - DO_NEXT_REFETCH_MAY_THROW
    - DO_NEXT_REFETCH_QUOTED

By organizing the evaluation to not use arbitrary index math on the
input series, this opens the doors to using sources for the data to
evaluate which are not random access (such as C vararg lists).  By
extending the EVAL settings for the evaluator, things that could
previously only be achieved in standard evaluation by functions like
APPLY/ONLY can now be done by both system code using the vararg API
*and* user code in Rebol.

To truly generalize the approach so that it can do arbitrary evaluation
from C vararg calls, as well as to deliver meaningful WHERE positions
during an error in a vararg argument evaluation, more work will be
needed to convert the varargs on-demand into an array.  Also, some
sort of caching will be needed in a debug mode to "see back in time"
to before the varargs had been advanced.

Though that hasn't been done yet, this lays the majority of the
groundwork for the method.  Generally speaking one shouldn't notice a
difference in the behavior (modulo bugs), but this is a very significant
rewrite--so expect speedbumps here and there with it in the near term.

One user-visible change, however, is the treatment of infix literal
function values (such as `do compose [1 (:+) 2]`).  See CureCode
issue #1934 for why this literal form of function does not behave as
infix--it must be bound to a word to have infix behavior.  This may
be temporary or permanent, but it is a change that is needed for at
least the current moment.