Take "do stack" down from series to pointer
When the initial implementation of Reb_Call as the fixed-size portion
of the stack frame was considered, it wasn't yet known how the frames
would be handled, in particular with cleanup of any resources that
were associated with the call.  It turned out stack frames would
not need to be dynamically allocated, and that the FRAME! elements
would be designed differently as REBSER nodes.

This meant the "TG_Do_Stack" series was not necessary, and the stack
frames could be walked with a simple linked list--where walking them
would only be relevant during their lifetime.  (A longjmp or return
would make them inaccessible.)