Simplify legacy true/none refinement/arg mode
When the legacy switch for enabling refinements was first added, it
seemed easy enough to just set LOGIC! in the unused refinement type
bits, and then have that indicate whether the function had "legacy"
expectations for the handling of refinements and refinement args.

In practice, that made the code much messier than just having a bit
on the function to tell whether it was legacy or not, because the
traversal of refinements led to losing the moment of the information.

This simplifies matters by picking one of the 8 type-specific EXT
bits for function and using it for this purpose.  Long term it would
be a waste of that bit, but only 2 are in use at the moment so there
is no rush...and it simplifies the evaluator code.