Check word for function before unset
Retrieved words are checked both to see if they are unset or function,
and if a function it is dispatched while unsets are errors.  Since the
function path has an action it can take, putting that first saves the
unset check on each word that turns out to be a function.