C++11 casting issues workaround
In the time since "Coherence One" where the C++11 and C++17 builds
were verified, there has been lots of new code and organization.  A
couple of uses of the cast macros are not approved by it, so this
temporarily turns them into C-style casts as a workaround until there
is time to focus on that issue.