Differentiate "Indexor" from integer index
This creates a separate REBIXR datatype, which is equivalent to REBCNT
in the C build but implemented as a class in the C++ build.  The name
"indexor" is short for "index or a flag" (although a trick is planned
which would use XOR of a number with 0 to count in inverted binary
for an alternate index form, so that would give it two reasons to
have that name.)

The class adds extra checking to ensure there are not accidental usages
of the "magic" states of THROWN_FLAG, END_FLAG, etc. such as trying
to increment them as if they were ordinary indices.  It also prevents
these states from being able to "leak" and find their way into places
that are not expecting very large numbers as magic values, but would
interpret them as literal array indices and crash.