Val_Init_Word_Unbound => Val_Init_Word
In the C source, there are almost no initializations of word values
with a binding at the time of explicit word creations.  Yet there are
many times that unbound words are created.

This swaps the bias so that Val_Init_Word is the version that isn't
bound, and if you want a binding you call Val_Init_Word_Bound()