Varargs by pointer dereference audit
Though the compiler doesn't complain, if you accidentally pass a pointer
to a va_arg list when it was expecting a va_arg list, you will get the
wrong output... sometimes.  Seems MinGW gives something different,
in any case.

In addition to correcting the dereferences, this chooses a varargs_ptr
naming scheme vs. varargs to stress in the mind of the reader that it
is a pointer, which will hopefully draw enough attention to wonder
"is this supposed to be dereferenced or not?" vs. just assuming.

Does some const correctness (and picks up a stray Val_Init_Word_Unbound
that somehow stuck around in a rebasing.)