Auto-reifying OBJECT! frames (breaks CLOSURE!)
This begins a unification of the way in which data is stored for a
given function invocation.  All cases start with just a chunk from the
chunk stack, and if there is a reason to refer to a stack frame via
an OBJECT! then one is created on demand.  When the stack frame goes
away, the object lives on but it is marked as not containing any data.
(Fields show as "--optimized out--" for starters, which is sort of a
stretch to put it that way, but it sounds good. :-P)

Because CLOSURE! had an intrinsic piece of functionality which was to
allow arguments to outlive the call, this breaks them.  The intended
solution in further commits is to hybridize objects so that they can
have fields that are "durable" vs. not.