Simplify context initialization protocol
Originally, the series behind OBJECT! was called a "frame" and had
a special REBVAL in its [0] slot called a FRAME! which could be used
to reconstruct an OBJECT! from just a pointer.  Other types were more
complex, and stored more fields in that 0 slot.  Ultimately these
complex types boiled down to just the additional ability to hold a
module spec.

Ren-C over time evolved this into storing a canon value of the
ANY-CONTEXT! itself, allowing any such value to be reconstituted from
a single pointer to its "varlist".  As it evolved, the responsibility
for filling this [0] value shifted to those building it...and the
Init routine became more and more a double-check, until it was just
an assert.

The assert was essentially checking redundant information, to the point
where to make the call to the assert the very data it was going to
be asserted against was extracted solely to pass the test.  This
commit removes that superfluous code, and also changes the invariant
on ANY-CONTEXT! to say that keylists are always managed--just to make
things a bit simpler.