Eliminate R_ARG1, R_ARG2, R_ARG3
While the fate of the R_RET return codes is not known (vs just having
natives always write to D_OUT), there is an argument that can be made
that `return R_TRUE;` is nicer than `SET_TRUE(D_OUT); return R_OUT;`

However, with the PARAM() and REFINE() macros, the ability to define
and refer to arguments by name instead of by number means that
`*D_OUT = *ARG(name); return R_OUT;` is an improvement over just
saying `return R_ARG2;` and having to go look to see what that is.
This doesn't completely review all instances to write them better, but
makes the move so that new references aren't introduced.

(Also updates a comment related to the native function dispatcher,
the absence of an observation causing lost time trying something that
motivated this change now... but turns out to not work.  Keeping this
since it was on the agenda to do anyway.)