Further simplify legacy refinement/arg processing
Now that there is a specific flag on functions to mark whether they
want legacy processing (refinements TRUE or NONE, args NONE) some of
the tricks to remember the state from a previous refinement are not
needed.  In particular, one trick used math on the arg slot, which is
a kind of "random access" mode the prefetch evaluator sought to forbid.

This simplifies the remaining code into asserts.