Prep END marker code trick for series info bits
The END marker trick (using the low bit of the header as 0 to signal
"not an end" and the second lowest bit as 0 to signal "not writable")
has gotten some field testing and having the nuances shaken out, so
this begins moving forward on use of the trick to terminate 1-element
value series resident inside a REBSER node.

This also makes the trick a bit more apparent by actually using the
Reb_Value_Header struct at the sites that are doing it.