Bring FRAME! into existence, adapt debug natives
This introduces the new ANY-CONTEXT! type of the FRAME!.  It also
cleans up the debugging natives, which do not need so many refinements
when there's a kind of object that can be queried directly for its
properties.

* Switches BACKTRACE to be like HELP in that it assumes you want a list
when it has no arguments in the console.  Eliminates many refinements
and returns a FRAME! by default when passed INTEGER! or FUNCTION!

* Added WHERE-OF, FUNCTION-OF, and LABEL-OF to reflect properties out
of a FRAME!

* Changed BREAKPOINT to be named PAUSE (after several times messing up
and typing BREAK, then having to igure out what the problem was.)

* Unify functions under common stack walking logic so that all can
take FUNCTION!, FRAME!, INTEGER!, NONE!