Move STACKVARS into context payload
This contains a number of changes surrounding the movement of the
pointer for stack variables into the payload of the context.  This
keeps it from overwriting the data pointer for the pooled allocation,
which long-term is beneficial for implementing a "hybrid" context
that has both stack allocated and pool-allocated ("<durable>") vars.

Getting this to work requried activating the code that allows a
series to contain a single value even without dynamic memory, because
the ANY-CONTEXT! value (the "[0] slot") has to be available even
in the absence of a heap allocation.  So this is a step toward the
necessary complexity to react to an OPT_SER_HAS_DYNAMIC flag being set
or not set.

The field in the context that was repurposed was the ".body" which was
something that was only used by TASK!.  Should something like TASK!
need to exist, it could contain a context pointer inside of it and
then have two other fields available.  Speaking in the hypothetical
future that TASK! were implemented in its intended form, the changes
to context would be beneficial because it would mean that the task
could have some local state that was "non-<durable>" and allocated on
the stack where it ran, to be freed when it finished.