Fix legacy arg1/arg2/arg3 error construction
A fairly simple hack makes an error archetype in legacy mode that adds
the arg1: arg2: arg3: fields that were required in every error for
R3-Alpha (but which are not required, and can be any names in Ren-C).

A combination of what appears to be a typo and a truncated length that
wasn't checked before (but is now checked with an assert) was breaking
error creation in `<r3-legacy>`.  This corrects both problems.