Distinguish THROWN_FLAG/CALL_MODE_THROW_PENDING
Historically, Rebol's do index (what Ren-C calls an "indexor") would
be set to THROWN_FLAG any time that a throw occurred.  This wiped out
the previous index number, which was good because it was information
that was no longer needed...by having it overwrite the index instead
of being a flag in some other place this avoided the accidental use of
a "bad" index position that could not be recovered.

Ren-C added new angles to this, including the "frameless" native.  A
frameless native does not use a nested Do_Core call or state, but
continues the evaluation inline.  So:

    if true [exit/from :if] print "this should print"

When a frameless IF is being processed, it uses the same index that
would be picked up to do the `print` when the IF is done.  There is
no new frame and no index that can be overwritten in that frame and
then get it back.  Hence if EXIT/FROM were to throw and overwrite
index as the thrown signal, the position before `print` would be lost.

An out-of-band flag is needed for these situations--which come up
whenever a frameless native is handling a throw that originates from
what would be one of its *arguments* (as opposed to handling a throw
during evaluation which would be one of its *parameters*).  This
parallels a situation inside the Do_Core evaluator itself which is
handled with a CALL_MODE of THROWN_PENDING.  So this just pushes that
responsibility to distinguish between `c->indexor = THROWN_FLAG` vs.
`c->mode = CALL_MODE_THROWN_PENDING` to the frameless native, which
is a subtlety but not an unmanageable one.

In the process this updates the Do_XXX_Core dispatchers to use the
same protocol to mutate the call state, vs return a result, so that
they can process the distinction in the same way.