Fix legacy OPTIONS_DONT_EXIT_NATIVES
The working of EXIT in Rebol2/R3-Alpha was equivalent to RETURN with
no argument...whereas in Ren-C it is a generic way of exiting any
function by frame, function value, or stack number.  A mode for keeping
the old EXIT behavior (exit the first user function or closure found
dynamically on the stack) had been extracted from Do_Core into the
stack level finding routine, and the legacy switch became entangled
with the stack numbering for backtrace.

This properly moves the special-case handling into REBNATIVE(exit).