Restore CLOSURE! behavior to persist args
Due to a rethinking of what order to do certain changes in, the move
to replace closure's "<durable> by default" with a more general notion
has been rescheduled to after solving some other problems.  Generally
speaking, the groundwork exists for "hybrid" frames that could have
some stack-based and some pool-based args/locals, and it's really
a matter of mapping the indices.

This also does a bit more shifting toward the hybrids by making it so
that the "stackvars" length is entirely separate from the "vars"
length.  A FRAME! as currently implemented thus has a vars of length
1 (which is done without dynamic allocation) and then stackvars length
according to the number of parameters to the function.  By eliminating
the need for a CONTEXT_VALUE() then 1 REBVAL is saved per call, and
the possibility exists to not push any stack data for functions which
have no parameters.