Store ANY-FUNCTION! as field in FRAME!
Building on the general work that allows an ANY-FUNCTION! to be saved
into just a single platform pointer, this makes the FRAME! category of
ANY-CONTEXT! use the slot otherwise allocated to "SPEC" (e.g. for
MODULE!) to save the function for the frame.  This allows it to be
both queried for it after the function call is terminated, and to
potentially contain all the information needed to invoke or re-invoke
a call.

Also finalizes the naming such that BOUND? is CONTEXT-OF