Specific binding in the debugger to FRAME!
This makes it so that when one is sitting at a stack frame in the
debugger, that the REPL will load the source as usual but then add in
an extra binding phase for the reified stack frame of the *specific*
instance of a function in question.  This means variables in recursive
functions can be examined with their value specific to that frame.

(The eventual hope is that this is how all functions work in their
bodies, if the mechanical finesse can be pulled off in a way that
beats the copy-and-rebind brute force of CLOSURE!)