Remove call frame label debug cache in GC
The label string of the word used to invoke a function (if any) is
helpful to see in the debugger when inspecting something having to
do with call frames.  Historically this was done on a per-function
basis where it might be interesting, but the debug build now caches
this label inside of the call itself so that anything that has a
Reb_Call in its hand can see it.

A superfluous instance of caching the label manually was also
accidentally dereferencing potentially uninitialized memory in the
debug build to do it.  Deleting it takes care of the problem--expand
the call frame in the debugger to see the name field where applicable.

Found by Valgrind.