Eliminate Reb_Call's "cell", use protecting pointers
This gets rid of a REBVAL-sized cell that was pushed in each call to
use for temporary storage of any value, that natives could take
advantage of by not having to GC protect it.  A motivation for its
existence was that a cell of memory was required by the implementation
of SET-PATH!'s logic--so as long as it was needed for that, it might
as well be offered for the advantage of a native.

The other REBVAL-sized value that was needed per call was a "safe"
from GC function value.  That was pared down to a single pointer using
a trick, and over time the other "cell" was seeming suspiciously not
necessary.  It was only used by a few natives, and in practice it was
possible to GC protect a value inexpensively by having cells be
in the C stack and just pointing the unused c->param and c->refine
pointers at it.

This modification makes Reb_Call a bit slimmer to the point of being
lightweight enough to consider using for things besides Do in order to
offer an interface to backtracing and FRAME! exposure.  Currently this
is speculated as a possible avenue for PARSE debugging.