Temporarily simplify data stack push
An inlining of the data stack pushing wasn't properly terminating, and
this was noticed by an aggressive GC debugging check.  The design of
the data stack is being reconsidered in general to perhaps be more
like the chunk stack and not an ordinary series...as contiguity is not
strictly a requirement for the operations it offers.

This simplifies the macro rather than fix it in anticipation of such
a change to the stack.