Fix R3-Alpha garbage trace on array input
When tracing a parse of a block, the debug string being used would
try and print it as a string.  This is because it was using the same
output format string as for string output (%s) instead of for a value
(%r).  Although there are better modes of tracing pending, it's easy
enough to fix.

However the idea of using numbered constants in this way for the
strings is actively harmful with no benefit over writing the strings
literally in the source.  If they are to be decoupled, there should
be some benefit to it.  Leaves a remark to that effect with the fix.