First step for PARSE to use FRAME!
This is a first step in the idea that PARSE use the same structure to
hold its state on recursions as a call frame for a function uses.  It
eliminates REBPARSE and uses Reb_Call instead.

The usage is not yet fully compatible with the debugger's expectations
of a fulfilled call frame.  Also, since not every recursion in PARSE
previously introduced a new REBPARSE, some of the state is "opaque"
still in local variables.  This makes a small step by having the
core parse hold its series and position being processed in a REBVAL
accessible as a stack variable, but has to save and restore some
variables which would be overwritten by the recursion.

As a first step in transformation, this is thorough and errs on the
side of caution.  But it should clean up as the PARSE code is vetted
to be hooked up to debugging.