Expression barriers: BAR! and LIT-BAR!
This introduces--and integrates--two new fundamental datatypes, the
BAR! and the LIT-BAR!.

(Note: a pending proposal would allow all types to have literal forms,
the so-called "lit-bit", in which this would add only one type.)

An expression barrier has a natural notation of a single vertical bar
`|`.  It leaves other words containing vertical bars alone, so `|a|`
and `a|b` are still WORD!.  It is inert in the evaluator except that
argument evaluation cannot accept a literal BAR! at the callsite.
That includes even "hard quoted" arguments, so `quote |` will cause
an error of hitting an expression barrier vs. quoting it.

However, if the caller can somehow generate a BAR! value, that will be
treated and type-checked normally.  LIT-BAR! is one way of doing it:

    >> append [a b c] '|
    == [a b c |]

However, any function that returns a BAR! works also.

Any number of BAR! will be consumed by a DO/NEXT:

    >> do/next [| | | | | 10 + 20] 'position
    == 30

From the user's point of view, PARSE behavior for `|` is unaffected.
It serves a different role in PARSE--as it could serve a different
role in any other dialect.

Relative to GROUP!, BAR! is both more lightweight visually and in
implementation--as a BAR! takes only one REBVAL-slot, and does not
require a recursion in the evaluator.  Dialects that generate code
can use it liberally and place it inline without needing to be
concerned with array management, and get expression separation
behavior.

The evaluative result of a BAR! is an UNSET!, so `do [a: 1 |]` will
be UNSET!.