Protect native frames by array, not keys
It would be dangerous if user code (in the debugger or otherwise) were
allowed to change the contents of stack variables used by natives.
The C code is not robust with respect to these variables changing
during a sub-evaluation, and could lead to problems that would crash
the interpreter.

The initial way to prevent the debugger from tampering with it was to
set all the keys in the parameter list to read-only.  However, this
meant that no frame could be created using the native's keylist, which
might be desirable to describe the parameters *prior* to an execution
(as opposed to modifying the out from underneath a running instance).

This moves the protection to be done at the moment of reification of
a running frame in the case that it is not a "user function" (FUNCTION!
or CLOSURE!)