Elaborate on END warning in type check
An assertion being hit frequently when things are not going right is
the check that happens during a request for a type of a value that
asserts that it does not have its low-bit clear...as that signals
an END.  However, this can often happen on garbage memory--because
half (or probably more than half) of arbitrary memory has the low
bit clear.

Saying it's unexpected and garbage thus may be more relevant in the
general case than mentioning it *could* be a signal of an END marker.
This updates the message and comment.