"Indexors", data stack pointers to native int
Rebol's general tradeoff is to cap series lengths at ~(2^32 / 2) and
use 32-bit integers to represent indexes and lengths.  This constraint
allows compatibility between the 32-bit and 64-bit versions.

Whether it's a good idea or not to have a 32-bit limit, using a 32-bit
*datatype* on a 64-bit platform hinders math and comparison,  This
changes two values that are frequently compared but not stored in the
system--the "indexor" used to traverse series during Do_Core, as well
as the type used to represent the "DSP" or "Data Stack Pointer".
Instead of using REBCNT and REBINT respectively, they use the C
`unsigned int` and `int` types respectively, which will be a size that
is most naturally optimized on the target system.