Remove `const` from REBSER instances
While const-correctness has been helpful on REBVAL*'s and character
or byte pointers, there has been no clear path for Rebol to effectively
use REBSER pointers as const or non-const.

Having to work around the constness for a few cases that are not adding
value is not worth it, so this removes const from the few REBSER-based
cases that were marked with it.