Test value flags w/o shifting, patch self bug
A REBVAL header historically contained two 8-bit sections for flags.
One was a set of "generic value flags" that applied to any value,
while the other was for type-specific flags...that each type could
interpret differently.

The way that flags were tested was that callsites had an integer that
they would pass of the flag they wanted, which was then shifted.  This
made it notationally more complex than if the flags were simply the
bits put to their right locations.  It also meant that there were two
routines for setting the flags.

This removes the need for the shifting and reduces the GET and SET
macros for the flags to just one set that can be used for both the
type-specific and general value flags.  It also adds safety in the
debug build by making the type-specific flags carry along an archetype
member of the type set as part of the mask, so that an alert will be
raised if one checks a typeset flag on a function (for instance).

In the process of doing this a bug was uncovered by which the SELF
word had been incorrectly getting bound, when it was "hidden" and thus
supposedly protected.  As fully solving the "definitional self"
question was not a focus of this commit, it works around the issue for
now by splitting "unbindability" out as a separate flag from "hidden".
PROTECT/HIDE sets both as before--but the SELF is only hidden.  Further
work on SELF is needed to fully explain its role in the system.

This also does some clarifying renaming, as VAL_GET_EXT and VAL_GET_OPT
were fairly opaque names, and flags like EXT_TYPESET_XXX raised
questions as well.  Also, OPT now meaning "optional" specifically w.r.t.
values was asking for trouble.  This picks the relatively obvious
names GET_VAL_FLAG and SET_VAL_FLAG, which are made easier and clearer
because they can set either category of flag on a value and do so with
a debug check as well.