Ferret out math usages of REB_XXX enum
In R3-Alpha, the datatype lived in the lowest bits of the header, which
meant VAL_TYPE() simply needed to mask the low bits to get the value
used in the type enum.  Ren-C's trick to eliminate the END! datatype
as well as allow a platform pointer-sized element to do "double duty"
shifted the type left two bits, in order to use the low bits for this
trick.  As a consequence, VAL_TYPE() required both masking and shifting
to the right two places.

As a frequently called operation (not just literal VAL_TYPE() but every
IS_XXX() macro calls VAL_TYPE()) the added shift is not completely
insignificant.  This commit switches the type enums to be in the range
of 0..255 instead of 0..63, spacing out the real values at every 4th
position.  As a consequence, it again only requires a mask to do a
type check with no shifting involved.

This threw a wrench into parts of the system that had the (bad)
assumption that math could be performed on the enum directly, such as
to say that REB_WORD + 1 was REB_SET_WORD.  Also, some routines wished
to index into an array that contained 64 elements, so they required
the shift.

So the majority of this commit is a system-wide audit to find these
specific usages and make them explicit.  There isn't much editorializing
at this point to decide if a usage that needs a shift or un-shift is
good or bad, although the IS_EVAL() macro is adapted to use a 256
element table instead of a 64-element one to avoid needing to shift
for that inner loop check.  (The "wasted" table entries could easily
be used for some other property of a type, by searching for the type
+ 1, so it is likely that multiple tables will merge together to
use that trick.)