Optimize stack indexing and expansion
R3-Alpha used an ordinary "block" style series of values to represent
its stack, in order to leverage the existing mechanics for expansion
and access operations.  However, it did not want to pay the cost of
checking on each PUSH to see if it was out of bounds of the series to
see if it needed to expand, nor did it maintain termination with an
"END!" value the way a normal series did.

So instead of using typical series append operations, it would bump
the stack by an arbitrary amount in those routines that were known to
run out of space.  Most were given headroom of 20 slots to work in,
with the assumption that if it were not enough then that routine
would have to include the manual bumping as well.

That was somewhat failure-prone, and Ren-C sought to have a stack
"service layer" that would be able to check for the need for expansion
on each push.  Rather than do this efficiently, it just reverted to
the append operations for typical series while stability was reached.

This commit brings back the optimization in a new form that permits any
push to potentially expand.  The stack is maintained as a block of
elements that are *not* END markers, terminated by one END marker.
As END markers are never pushed to the stack, need for expansion can
be determined by bumping the data stack pointer and seeing if it's
an END or not...if not, the slot is safe to use.  Dropping the stack
will always leave a non-END behind, so that is legal too.

The garbage collector must handle this series specially, because though
it has "non-END" values going from after the data stack pointer to the
end of the array, these values should not be GC protected (as they
have been popped or dropped).  Consequently they are also potentially
invalid.  The GC must thus temporarily put an END marker in the data
stack after the DSP, and then restore it to a non-END marker (assuming
the DSP wasn't at the very end already).