Deprecate CLOSURE => <durable>, PROC, PROCEDURE
This commit makes the next step toward FUNCTION! and CLOSURE!
unification by eliminating the CLOSURE! type, in favor of the ability
to mark a function's arguments <durable>.

(Note: The name <durable> is tentative, and at present durable is "all
or nothing"...it is not possible to specify some arguments or locals
as surviving and others not.  However, that selectivity is planned in
the design and should be added at a later date.)

Marking a function's arguments <durable> currently has the other effect
historically of a closure--to give them specific binding.  This feature
is planned to be implemented in a way that does not require a deep
copy of the function's body on each execution as closure does--so all
FUNCTION!s will have that property.  Only the selection of individual
arguments as durable will remain.

Taking the place of CLOS and CLOSURE are the PROC and PROCEDURE
function generators.  Instead of RETURN they have a definitional-scoped
0-arity function called LEAVE.  This returns UNSET! always, and in
addition they suppress any return value from falling out:

     >> foo: procedure [] [leave print "This won't print"]
     >> foo
     ; No result

     >> foo: proc [y] [y + 20]
     >> foo 10
     ; No result