Shorten accessor prefixes: ARR_, SER_, CTX_
This is a search-and-replace on the prefixes ARRAY_, SERIES_, and
CONTEXT_ to shortened forms.  The replacement is only on the prefixes,
to line up with the accessor macros VAL_XXX.

This is actually a step back closer to R3-Alpha.  For instance, the
"array" accessors were in terms of "BLK", so BLK_HEAD and BLK_AT etc.
When first moving toward the change to distinguish the "array" category
from the BLOCK! data type, it wasn't clear if the tradeoff of clarity
vs. length would be worth it to abbreviate--hence the expansion.

In the balance of usage, the necessity of writing out the entire name
is often distracting.  For instance, `CONTEXT_KEYS_HEAD(context)` is
more wordy than `CTX_KEYS_HEAD(context)`, and one is usually cued
already by the name of the variable to know what is being operated on--
hence emphasizing the extracted property is more pertinent.

When not a prefix, the words remain written out.  So `ARRAY_SERIES` to
extract the series from an array becomes `ARR_SERIES` and not `ARR_SER`
for instance.

This also makes the choice to switch from REBCON for contexts to
REBCTX.  Though X is a somewhat "disruptive" character, the convention
in programming to understand CTX as "context" and CON as things like
"console" or "control" is fairly entrenched.  Not enough benefit
emerges from resisting that interpretation to be worth it when it's
used as a prefix in this way.

While most things are shortened by this commit, some got longer...in
part in order to distinguish them.  e.g. SERIES_FLAG_XXX stands out
as not being one of the SER_XXX routines, and hence does not take a
parameter but is a constant.