Reintroduce ET_XXX eval types, DO tidying
This brings back an R3-Alpha idea of turning the REB_XXX types into an
ET_ evaluation type enumeration prior to doing a switch() statement on
them.  Ren-C initially was aiming to make the datatypes in types.r
which required evaluation contiguous, so that they would lead to
a jump table optimization and that the default would fall through to
be the "no evaluation behavior" class.  Yet that reorganization would
tie the hands of the system to use the numbers in other ways, and now
the REB_XXX types skip by 4 so the types range 0..63 to 0..255.

So this just puts the idea back with a new twist of ET_NONE being 0 to
let the ANY_EVAL() macro use the same trick.

This also tidies up some repetitive code related to the definitional
RETURN and LEAVE processing, and begins the process of paring down the
comments and asserts to build Do_Core() into coherent sections with
notes reduced into a "developer's guide" or issue database tickets.